test: move cpfValidator to validation layer
