+++ src/presentation/controllers/load-person/load-person-controller.spec.ts
@@ -1,6 +1,7 @@
-import { type CpfValidator, type PersonModel, type LoadPerson } from './load-person-protocols'
+import { type PersonModel, type LoadPerson } from './load-person-protocols'
 import { InvalidParamError, NotFoundError, ServerError } from '../../errors'
 import { LoadPersonController } from './load-person-controller'
+import { type CpfValidator } from '../../../validation/protocols'
 
 interface SutTypes {
   sut: LoadPersonController
